+++ src/main/kotlin/com/mmm/internal/data/client/_TiktokApiClient.kt
@@ -1,0 +1,50 @@
+package com.mmm.internal.data.client
+
+import com.mmm.internal.common.dsl.WebClientBuilder
+import com.mmm.internal.common.util.mapToTrendingResponse
+import com.mmm.internal.data.client.response.TiktokTrendingResponse
+import com.mmm.internal.data.property.TiktokApiProperties
+import org.springframework.beans.factory.annotation.Qualifier
+import org.springframework.stereotype.Component
+import org.springframework.web.reactive.function.client.WebClient
+import reactor.core.publisher.Mono
+import reactor.core.scheduler.Schedulers
+import java.time.Duration
+
+/**
+ *
+ */
+@Component
+private class _TiktokApiClient(
+    @Qualifier("tiktok") private val tiktokClient: WebClient,
+    private val tiktokApiProperties: TiktokApiProperties
+): TiktokApiClient {
+
+    private val webClientBuilder: WebClientBuilder = WebClientBuilder()
+
+    override fun trending(
+        count: Int
+    ): Mono<List<TiktokTrendingResponse>> {
+        if (count > 3) {
+            return Mono.empty()
+        }
+
+        return webClientBuilder.webClient(tiktokClient)
+            .uri(tiktokApiProperties.trendingApi)
+            .queryParam("count", count)
+            .queryParam("region", REGION)
+            .queryParam("language", LANGUAGE)
+            .queryParam("tz_name", TIMEZONE)
+            .build { it.mapToTrendingResponse() }
+            .subscribeOn(Schedulers.boundedElastic())
+            .timeout(Duration.ofSeconds(10))
+            .onErrorResume { Mono.empty() }
+    }
+
+    companion object {
+        private const val REGION = "KR"
+        private const val LANGUAGE = "ko"
+        private const val TIMEZONE = "Asia%2FSeoul"
+    }
+
+}
